backend refactro done (#411)
