+++ controllers/erpControllers/mailInvoiceController.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const custom = require('../corsControllers/custom');
+const custom = require('@/controllers/middlewaresControllers/pdfController');
 const { SendInvoice } = require('../../emailTemplate/SendInvoice');
 const mongoose = require('mongoose');
 const InvoiceModel = mongoose.model('Invoice');
