+++ controllers/middlewaresControllers/createCRUDController/listAll.js
@@ -1,0 +1,34 @@
+const listAll = async (Model, req, res) => {
+  const limit = parseInt(req.query.limit) || 100;
+  const sort = parseInt(req.query.sort) || 'desc';
+  try {
+    //  Query the database for a list of all results
+    const result = await Model.find({ removed: false })
+      .limit(limit)
+      .sort({ created: sort })
+      .populate();
+
+    if (count > 0) {
+      return res.status(200).json({
+        success: true,
+        result,
+        message: 'Successfully found all documents',
+      });
+    } else {
+      return res.status(203).json({
+        success: false,
+        result: [],
+        message: 'Collection is Empty',
+      });
+    }
+  } catch (err) {
+    return res.status(500).json({
+      success: false,
+      result: [],
+      message: 'Oops there is an Error',
+      error: err,
+    });
+  }
+};
+
+module.exports = listAll;
