+++ controllers/middlewaresControllers/createCRUDController/read.js
@@ -1,0 +1,31 @@
+const read = async (Model, req, res) => {
+  try {
+    // Find document by id
+    const result = await Model.findOne({ _id: req.params.id, removed: false });
+    // If no results found, return document not found
+    if (!result) {
+      return res.status(404).json({
+        success: false,
+        result: null,
+        message: 'No document found by this id: ' + req.params.id,
+      });
+    } else {
+      // Return success resposne
+      return res.status(200).json({
+        success: true,
+        result,
+        message: 'we found this document by this id: ' + req.params.id,
+      });
+    }
+  } catch (err) {
+    // Server Error
+    return res.status(500).json({
+      success: false,
+      result: null,
+      message: 'Oops there is an Error',
+      error: err,
+    });
+  }
+};
+
+module.exports = read;
