+++ controllers/middlewaresControllers/createCRUDController/remove.js
@@ -1,0 +1,39 @@
+const remove = async (Model, req, res) => {
+  try {
+    // Find the document by id and delete it
+    let updates = {
+      removed: true,
+    };
+    // Find the document by id and delete it
+    const result = await Model.findOneAndUpdate(
+      { _id: req.params.id, removed: false },
+      { $set: updates },
+      {
+        new: true, // return the new result instead of the old one
+      }
+    ).exec();
+    // If no results found, return document not found
+    if (!result) {
+      return res.status(404).json({
+        success: false,
+        result: null,
+        message: 'No document found by this id: ' + req.params.id,
+      });
+    } else {
+      return res.status(200).json({
+        success: true,
+        result,
+        message: 'Successfully Deleted the document by id: ' + req.params.id,
+      });
+    }
+  } catch (err) {
+    return res.status(500).json({
+      success: false,
+      result: null,
+      message: 'Oops there is an Error',
+      error: err,
+    });
+  }
+};
+
+module.exports = remove;
