+++ controllers/middlewaresControllers/createCRUDController/update.js
@@ -1,0 +1,42 @@
+const update = async (Model, req, res) => {
+  try {
+    // Find document by id and updates with the required fields
+    const result = await Model.findOneAndUpdate({ _id: req.params.id, removed: false }, req.body, {
+      new: true, // return the new result instead of the old one
+      runValidators: true,
+    }).exec();
+    if (!result) {
+      return res.status(404).json({
+        success: false,
+        result: null,
+        message: 'No document found by this id: ' + req.params.id,
+      });
+    } else {
+      return res.status(200).json({
+        success: true,
+        result,
+        message: 'we update this document by this id: ' + req.params.id,
+      });
+    }
+  } catch (err) {
+    // If err is thrown by Mongoose due to required validations
+    if (err.name == 'ValidationError') {
+      return res.status(400).json({
+        success: false,
+        result: null,
+        message: 'Required fields are not supplied',
+        error: err,
+      });
+    } else {
+      // Server Error
+      return res.status(500).json({
+        success: false,
+        result: null,
+        message: 'Oops there is an Error',
+        error: err,
+      });
+    }
+  }
+};
+
+module.exports = update;
