+++ routes/erpRoutes/erpApi.js
@@ -1,13 +1,8 @@
 const express = require('express');
-const multer = require('multer');
-const path = require('path');
-const setFilePathToBody = require('@/middlewares/setFilePathToBody');
+
 const { catchErrors } = require('@/handlers/errorHandlers');
 
 const router = express.Router();
-
-const adminController = require('@/controllers/erpControllers/adminController');
-const roleController = require('@/controllers/erpControllers/roleController');
 
 const employeeController = require('@/controllers/erpControllers/employeeController');
 const paymentModeController = require('@/controllers/erpControllers/paymentModeController');
@@ -22,48 +17,6 @@ const expenseCategoryController = require('@/controllers/erpControllers/expenseC
 const paymentInvoiceController = require('@/controllers/erpControllers/paymentInvoiceController');
 const settingsController = require('@/controllers/erpControllers/settingsController');
 const offerController = require('@/controllers/erpControllers/offerController');
-
-// //_______________________________ Admin management_______________________________
-
-var adminPhotoStorage = multer.diskStorage({
-  destination: function (req, file, cb) {
-    cb(null, 'public/uploads/admin');
-  },
-  filename: function (req, file, cb) {
-    cb(null, Date.now() + path.extname(file.originalname));
-  },
-});
-const adminPhotoUpload = multer({ storage: adminPhotoStorage });
-
-router
-  .route('/admin/create')
-  .post([adminPhotoUpload.single('photo'), setFilePathToBody], catchErrors(adminController.create));
-router.route('/admin/read/:id').get(catchErrors(adminController.read));
-// router.route("/admin/update/:id").patch(catchErrors(adminController.update));
-// router.route("/admin/delete/:id").delete(catchErrors(adminController.delete));
-router.route('/admin/search').get(catchErrors(adminController.search));
-router.route('/admin/list').get(catchErrors(adminController.list));
-router.route('/admin/profile').get(catchErrors(adminController.profile));
-router.route('/admin/status/:id').patch(catchErrors(adminController.status));
-// router
-//   .route("/admin/photo")
-//   .post(
-//     [adminPhotoUpload.single("photo"), setFilePathToBody],
-//     catchErrors(adminController.photo)
-//   );
-// router
-//   .route("/admin/password-update/:id")
-//   .patch(catchErrors(adminController.updatePassword));
-
-// //____________________________ Role management_______________________________
-
-router.route('/role/create').post(catchErrors(roleController.create));
-router.route('/role/read/:id').get(catchErrors(roleController.read));
-router.route('/role/update/:id').patch(catchErrors(roleController.update));
-router.route('/role/delete/:id').delete(catchErrors(roleController.delete));
-router.route('/role/search').get(catchErrors(roleController.search));
-router.route('/role/list').get(catchErrors(roleController.list));
-router.route('/role/filter').get(catchErrors(roleController.filter));
 
 // //_________________________________ API for employees_____________________
 router.route('/employee/create').post(catchErrors(employeeController.create));
