+++ setup/reset.js
@@ -5,11 +5,9 @@ mongoose.connect(process.env.DATABASE);
 mongoose.Promise = global.Promise; // Tell Mongoose to use ES6 promises
 
 async function deleteData() {
-  const Admin = require('../models/erpModels/Admin');
+  const Admin = require('../models/coreModels/Admin');
   await Admin.remove();
-  console.log(
-    '👍👍👍👍👍👍👍👍 admin Deleted. To setup demo admin data, run\n\n\t npm run setup\n\n'
-  );
+  console.log('👍 admin Deleted. To setup demo admin data, run\n\n\t npm run setup\n\n');
   process.exit();
 }
