+++ setup/setup.js
@@ -6,7 +6,7 @@ mongoose.Promise = global.Promise; // Tell Mongoose to use ES6 promises
 
 async function createAdmin() {
   try {
-    const Admin = require('../models/erpModels/Admin');
+    const Admin = require('../models/coreModels/Admin');
     var newAdmin = new Admin();
     const passwordHash = newAdmin.generateHash('admin123');
 
@@ -16,10 +16,10 @@ async function createAdmin() {
       name: 'Salah Eddine',
       surname: 'Lalami',
     }).save();
-    console.log('👍👍👍👍👍👍👍👍 Admin created : Done!');
+    console.log('👍 Admin created : Done!');
     process.exit();
   } catch (e) {
-    console.log('\n👎👎👎👎👎👎👎👎 Error! The Error info is below');
+    console.log('\n🚫 Error! The Error info is below');
     console.log(e);
     process.exit();
   }
